requirements.txt
ultralytics
easyocr
paddleocr
pytesseract
torch
torchvision
transformers
pillow
opencv-python
icrawler
numpy
kaggle
pyyaml
matplotlib
streamlit
